Create MoveAndLink function and Update errors for Linking logic and MoveFile function
Separate the Linking logic to a Link function.
The new MoveAndLink function handles moving the files, renaming the
directories and then calls the Link Function
This allows for the special case of Dest file existing and source file
not existing to be handle by calling the Link function with reversed
arguments.
This Should work theoretically, Needs testing.