Move everything to packages again. Make every command into methods of an Application struct. And Call functions with fields of struct
Move the Run() logic to linksym.go and Create an Application struct with
all subcommands as receivers on it.
Split everything into 5 packages - Flags, Logger, Commands, Link, Config
Remove the Global vars InitDir, HomeDir, and ConfigPath in favor of
struct fields.
All AppConfig methods are called with App.Configuration
Rename file.go to Path.go and move AliasPath and ExpandPath to it