(Fix): Add command special case of "Not needing the move step" no longer requires user to add a Trailing slash to indicate if source is a directory or file
Previously when making a link to a file or directory that had already
been moved by the user, required the path to have a trailing slash to
indicate if the soure is a directory or file to link to corresponding
file or directory. Which is longer needed, since the MoveAndLink()
function is not called for moving directories.

Also update Error messages to use aliased paths for source and
destination