Fix: issue with linking of 2 files both existing should be fixed.
Not sure if delete the source file and then creating the symlink in it's
place is the best solution. or What bugs and edge cases this might
create.
When creating symlinks using linksym source, if files from dotfiles
directory were to be linked, and the file existed on the sources, it
would create errors